requirements.txt
streamlit==1.22.0
pillow==11.3.0
numpy==2.0.2
tensorflow==2.9.0
openai==0.28.1
protobuf==3.20.3
